Adds policies app
