+++ app/grandchallenge/core/models.py
@@ -1,7 +1,6 @@
 import uuid
 
 from django.conf import settings
-from django.core.exceptions import ValidationError
 from django.db import models
 
 
@@ -72,12 +71,3 @@ class RequestBase(models.Model):
 
     class Meta:
         abstract = True
-
-
-class TermsOfService(models.Model):
-    body = models.TextField()
-
-    def save(self, *args, **kwrags):
-        if not self.pk and TermsOfService.objects.exists():
-            raise ValidationError("Only one terms of service instance allowed")
-        return super().save(*args, **kwrags)
