+++ app/grandchallenge/policies/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from grandchallenge.policies.models import TermsOfService
+
+admin.site.register(TermsOfService)
