+++ app/tests/factories.py
@@ -13,12 +13,12 @@ from grandchallenge.challenges.models import (
     ExternalChallenge,
     ImagingModality,
 )
-from grandchallenge.core.models import TermsOfService
 from grandchallenge.datasets.models import AnnotationSet, ImageSet
 from grandchallenge.evaluation.models import Job, Method, Result, Submission
 from grandchallenge.jqfileupload.models import StagedFile
 from grandchallenge.pages.models import Page
 from grandchallenge.participants.models import RegistrationRequest
+from grandchallenge.policies.models import TermsOfService
 from grandchallenge.teams.models import Team, TeamMember
 from grandchallenge.uploads.models import UploadModel
 from grandchallenge.workstation_configs.models import WorkstationConfig
